replace `SegyFileAsync` with `SegyFileWrapper` and refactor SEG-Y file utilities
